requirements.txt
langchain
streamlit
google-generativeai
python-dotenv
chromadb
faiss-cpu
langchain-openai
pathlib
pypdf